tests/unit/web/routers/test_audit_logs.py
"""
Unit tests for the audit logs router.

This module contains tests for the audit logs-related API endpoints, focusing on 
request validation, response formatting, and permission checks.
"""

import pytest
from unittest.mock import patch, MagicMock
from fastapi import HTTPException
from fastapi.testclient import TestClient

from netraven.web.routers.audit_logs import router
from netraven.web.auth import UserPrincipal
from netraven.web.schemas.audit_log import AuditLogList, AuditLogResponse


@pytest.fixture
def audit_log_service_mock():
    """Fixture that returns a mock audit logging service."""
    with patch("netraven.web.routers.audit_logs.AuditService") as mock:
        yield mock


@pytest.fixture
def audit_log_fixture():
    """Fixture that returns test audit log data."""
    return {
        "id": "test-audit-id-123",
        "event_type": "auth",
        "event_name": "login",
        "actor_id": "test-user",
        "actor_type": "user",
        "target_id": None,
        "target_type": None,
        "ip_address": "127.0.0.1",
        "user_agent": "Mozilla/5.0",
        "session_id": None,
        "description": "User test-user logged in",
        "status": "success",
        "event_metadata": {"source": "web"},
        "created_at": "2023-01-01T00:00:00"
    }


@pytest.fixture
def admin_principal():
    """Fixture that returns a mock admin principal."""
    principal = MagicMock(spec=UserPrincipal)
    principal.username = "admin"
    principal.is_admin = True
    principal.has_scope.return_value = True
    return principal


class TestAuditLogsRouter:
    """Test suite for the audit logs router."""

    def test_list_audit_logs(self, audit_log_service_mock, audit_log_fixture, admin_principal):
        """Test listing audit logs."""
        # Setup mock responses
        audit_log_service_mock.get_audit_logs.return_value = {
            "items": [audit_log_fixture],
            "total": 1
        }
        
        # TODO: Implement actual test when audit endpoint is created
        pass
        
    def test_list_audit_logs_unauthorized(self, audit_log_service_mock):
        """Test that listing audit logs requires proper permissions."""
        # Setup mock principal without admin permissions
        non_admin_principal = MagicMock(spec=UserPrincipal)
        non_admin_principal.username = "regular_user"
        non_admin_principal.is_admin = False
        non_admin_principal.has_scope.return_value = False
        
        # TODO: Implement actual test when audit endpoint is created
        pass
        
    def test_get_audit_log_by_id(self, audit_log_service_mock, audit_log_fixture, admin_principal):
        """Test getting a specific audit log entry."""
        # Setup mock response
        audit_log_service_mock.get_audit_log_by_id.return_value = audit_log_fixture
        
        # TODO: Implement actual test when endpoint is created
        pass
        
    def test_get_audit_log_by_id_not_found(self, audit_log_service_mock, admin_principal):
        """Test requesting a non-existent audit log entry."""
        # Setup mock response for not found
        audit_log_service_mock.get_audit_log_by_id.return_value = None
        
        # TODO: Implement actual test when endpoint is created
        pass 