requirements.txt
# Core dependencies
fastapi==0.100.1
uvicorn==0.20.0
flask==2.2.2
gunicorn==20.1.0
sqlalchemy==2.0.7
pydantic==2.4.2
email-validator==1.3.1
werkzeug==2.2.3

# Authentication
python-jose==3.3.0
pyjwt==2.6.0
python-multipart==0.0.5
bcrypt==4.0.1
passlib==1.7.4
cryptography==41.0.3

# Database
psycopg2-binary==2.9.5
alembic==1.10.3

# Utilities
requests==2.28.2
prometheus-client==0.16.0
schedule==1.1.0
PyYAML==6.0
